test: Fix dead loop on *mb_rand_update_test
When job finished, the tests will fetch next job with
index plus one. And it does not check the job context
status. If the next job is processing or finished status,
data will not be updated and the digest is wrong.

This patch will check the status of job , only IDLE job
can be update.

Change-Id: Ia9e72267926a12d912c00ed3676a2a65bba18bb1
Signed-off-by: Jerry Yu <jerry.h.yu@arm.com>