aes_gcm: Port of single-buffer vaes version
Ported changes to AES-GCM for vaesni avx512 instructions from the Intel IPSec
multi-buffer library. Although the single-buffer aes-gcm was originally from
ISA-L, the ipsec imported for when multi-buffer was not possible and made a
number of updates. Attempted to keep as similar as possible to ease future
porting between the libraries.

Change-Id: If601c5a63ddfb824f0436b72999076d63ffdcb2d
Signed-off-by: Greg Tucker <greg.b.tucker@intel.com>