test: Fix hash reference functions for gcc11 issue
Starting with gcc11, optimization is causing the hash reference
functions to fail on O2 but not on O1. Likely a gcc middle-end
bug (https://gcc.gnu.org/bugzilla/show_bug.cgi?id=101257) that
causes violation of strict-aliasing rules. Until a work around
is found, this uses an attribute to roll back ipa.

Change-Id: Ie525bb890f59b6a1ab83f42e3951a4b85f45bd8c
Signed-off-by: Greg Tucker <greg.b.tucker@intel.com>